Refactor project structure and namespaces
This commit reorganizes the project structure by replacing the `Data.Interface` namespace with `Application.Interface.Persistence`, promoting a more application-centric architecture. New project references have been added to enhance access to domain models. Several interfaces and application classes have been updated to reflect the new namespace organization, ensuring consistency across the codebase. Additionally, new projects for `Persistence.Data` and `Persistence.Repository` have been introduced, while references to the old `Data.Interface` and `Data.Repository` projects have been removed to streamline dependencies. Overall, these changes improve the modularity and maintainability of the codebase.
