+++ MarCorp.DemoBack.Application.Interface/MarCorp.DemoBack.Application.Interface.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MarCorp.DemoBack.Application.DTO\MarCorp.DemoBack.Application.DTO.csproj" />
+    <ProjectReference Include="..\MarCorp.DemoBack.Domain.Models\MarCorp.DemoBack.Domain.Models.csproj" />
     
     <ProjectReference Include="..\MarCorp.DemoBack.Support.Common\MarCorp.DemoBack.Support.Common.csproj" />
   </ItemGroup>
