+++ MarCorp.DemoBack.Application.Interface/Persistence/IDapperContext.cs
@@ -1,6 +1,6 @@
 ﻿using System.Data;
 
-namespace MarCorp.DemoBack.Data.Interface
+namespace MarCorp.DemoBack.Application.Interface.Persistence
 {
     public interface IDapperContext
     {
