+++ MarCorp.DemoBack.Application.Interface/UseCases/IUsersApplication.cs
@@ -1,7 +1,7 @@
 ﻿using MarCorp.DemoBack.Application.DTO;
 using MarCorp.DemoBack.Support.Common;
 
-namespace MarCorp.DemoBack.Application.Interface
+namespace MarCorp.DemoBack.Application.Interface.UseCases
 {
     public interface IUsersApplication
     {
