+++ MarCorp.DemoBack.Application.Main/MarCorp.DemoBack.Application.Main.csproj
@@ -14,7 +14,6 @@
   <ItemGroup>
     <ProjectReference Include="..\MarCorp.DemoBack.Application.Interface\MarCorp.DemoBack.Application.Interface.csproj" />
     <ProjectReference Include="..\MarCorp.DemoBack.Application.Validator\MarCorp.DemoBack.Application.Validator.csproj" />
-    <ProjectReference Include="..\MarCorp.DemoBack.Domain.Interface\MarCorp.DemoBack.Domain.Interface.csproj" />
     <ProjectReference Include="..\MarCorp.DemoBack.Domain.Models\MarCorp.DemoBack.Domain.Models.csproj" />
     <ProjectReference Include="..\MarCorp.DemoBack.Support.Common\MarCorp.DemoBack.Support.Common.csproj" />
   </ItemGroup>
