+++ MarCorp.DemoBack.Application.Main/UsersApplication.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using FluentValidation;
 using MarCorp.DemoBack.Application.DTO;
-using MarCorp.DemoBack.Application.Interface;
+using MarCorp.DemoBack.Application.Interface.UseCases;
 using MarCorp.DemoBack.Domain.Interface;
 using MarCorp.DemoBack.Support.Common;
