+++ MarCorp.DemoBack.Application.Test/UsersApplicationTest.cs
@@ -1,4 +1,4 @@
-﻿using MarCorp.DemoBack.Application.Interface;
+﻿using MarCorp.DemoBack.Application.Interface.UseCases;
 using MarCorp.DemoBack.Application.Main;
 using MarCorp.DemoBack.Application.Validator;
 using MarCorp.DemoBack.Data.Connections;
