--- MarCorp.DemoBack.Data.Repository/MarCorp.DemoBack.Data.Repository.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Dapper" Version="2.1.35" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MarCorp.DemoBack.Data.Connections\MarCorp.DemoBack.Data.Connections.csproj" />
-    <ProjectReference Include="..\MarCorp.DemoBack.Data.Interface\MarCorp.DemoBack.Data.Interface.csproj" />
-  </ItemGroup>
-
-</Project>
