+++ MarCorp.DemoBack.Persistence.Repository/MarCorp.DemoBack.Persistence.Repository.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Dapper" Version="2.1.35" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MarCorp.DemoBack.Application.Interface\MarCorp.DemoBack.Application.Interface.csproj" />
+    <ProjectReference Include="..\MarCorp.DemoBack.Data.Connections\MarCorp.DemoBack.Persistence.Data.csproj" />
+    <ProjectReference Include="..\MarCorp.DemoBack.Persistence.Data\MarCorp.DemoBack.Persistence.Data.csproj" />
+  </ItemGroup>
+
+</Project>
